extract read_result
