+++ src/blob/index/mod.rs
@@ -1,4 +1,7 @@
-use crate::{prelude::*, storage::ReadResult};
+use crate::{
+    prelude::*,
+    storage::{BlobRecordTimestamp, ReadResult},
+};
 
 mod bptree;
 mod core;
