+++ src/blob/mod.rs
@@ -4,7 +4,6 @@ mod file;
 pub(crate) mod header;
 pub(crate) mod index;
 
-pub use self::core::BlobRecordTimestamp;
 pub(crate) use self::core::BLOB_INDEX_FILE_EXTENSION;
 pub(crate) use self::core::{Blob, FileName};
 pub use self::entry::Entry;
